fix: return playback function after it reaches EOF
The program would run indefinitely after EOF was reached. This is fixed
by toggling a global variable and checking for it in the playback
function.
The decoder and device should not be unitialized inside the data
callback
The global variable is toggled to false after inside the if block to
allow the same behevaior when playing multiple songs